pyproject.toml
[tool.poetry]
name = "carbonlens"
version = "0.1.0"
description = "Personal CO2 tracker & climate coach"
authors = ["Team CarbonLens"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {version="^0.30.6", extras=["standard"]}
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
SQLAlchemy = "^2.0.36"
alembic = "^1.13.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
PyYAML = "^6.0.2"
pandas = "^2.2.3"
numpy = "^2.1.2"
scikit-learn = "^1.5.2"
prophet = "^1.1.5"
plotly = "^5.24.1"
streamlit = "^1.39.0"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
black = "^24.10.0"
ruff = "^0.6.8"
mypy = "^1.13.0"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
